requirements.txt
alembic==1.16.1
attrs==25.3.0
blinker==1.9.0
certifi==2025.6.15
charset-normalizer==2.1.1
click==8.2.1
Faker==14.2.0
Flask==2.2.2
flask-cors==6.0.1
Flask-JWT-Extended==4.7.1
Flask-Mail==0.10.0
Flask-Migrate==3.1.0
Flask-SQLAlchemy==3.0.3
greenlet==3.2.3
gunicorn==23.0.0
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.6
Mako==1.3.10
MarkupSafe==3.0.2
packaging==25.0
psycopg2-binary==2.9.9
PyJWT==2.10.1
PyMySQL==1.1.1
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
pytz==2024.2
requests==2.28.1
SQLAlchemy==2.0.29
sqlalchemy-serializer==1.4.22
tomli==2.2.1
typing_extensions==4.14.0
tzdata==2025.2
urllib3==1.26.20
Werkzeug==2.2.2
